fixed post serializer to include author
