+++ frontend/src/App.jsx
@@ -7,7 +7,7 @@ import Home from "./components/Home";
 import Signup from "./components/signup/Signup";
 import Login from "./components/login/Login";
 import Stream from "./components/stream/Stream";
-import ProfilePage from "./components/profilePage/ProfilePage";
+import Profile from "./components/profile/Profile";
 import NavigationBar from "./components/NavigationBar";
 
 import axios from "axios";
@@ -24,7 +24,7 @@ class App extends Component {
             <Route path="/signup" component={Signup} />
             <Route path="/login" component={Login} />
             <Route path="/stream" component={Stream} />
-            <Route path="/author/:id" component={ProfilePage} />
+            <Route path="/author/:id" component={Profile} />
             <Route exact path="/" component={Home} />
           </Switch>
         </Root>
