+++ frontend/src/Reducer.js
@@ -4,7 +4,7 @@ import { connectRouter } from "connected-react-router";
 import { signupReducer } from "./components/signup/SignupReducer";
 import { loginReducer } from "./components/login/LoginReducer";
 import { streamReducer } from "./components/stream/StreamReducer";
-import { profileReducer } from "./components/profilePage/ProfileReducer";
+import { profileReducer } from "./components/profile/ProfileReducer";
 
 const createRootReducer = (history) =>
   combineReducers({
