+++ frontend/src/components/profile/Divider.jsx
