--- frontend/src/components/profilePage/ProfileIcon.jsx
@@ -1,20 +1,0 @@
-/*
-SVG components:
-    ProfileIcon
-*/
-
-function ProfileIcon(props) {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width={props.svgScale} height="auto" viewBox="0 0 130.764 130.764">
-        <g id="Group_47" data-name="Group 47" transform="translate(-804.583 -705.414)">
-            <path id="Ellipse_23" d="M65.352,0a65.4,65.4,0,1,0,.03,0Zm.031,23.889c14.251,0,16.707,9.085,16.707,20.293S74.61,64.475,65.383,64.475,48.676,55.39,48.676,44.182,51.132,23.889,65.383,23.889ZM56.234,66.031s2.746,3.5,9.148,3.5,9.149-3.5,9.149-3.5C94.043,69.553,96.593,75.117,96.9,94.641a12.572,12.572,0,0,0,.031,1.354v1.518s-4.644,9.365-31.553,9.365-31.555-9.365-31.555-9.365V95.461c.005.18.018.1.047-1.455C34.233,75,36.934,69.514,56.234,66.031Z" transform="translate(804.583 705.414)"/>
-            <g id="_74472" data-name="74472" transform="translate(838.406 729.303)">
-            <path id="Path_125" data-name="Path 125" class="cls-1" d="M41.9,300.723C41.9,300.04,41.894,300.53,41.9,300.723Z" transform="translate(-41.895 -229.152)"/>
-            <path id="Path_126" data-name="Path 126" class="cls-1" d="M308.085,302.107C308.094,301.92,308.085,300.807,308.085,302.107Z" transform="translate(-244.97 -230.002)"/>
-            </g>
-        </g>
-        </svg>
-    );
-}
-
-export { ProfileIcon };
