--- frontend/src/components/profilePage/styling.jsx
@@ -1,12 +1,0 @@
-var color = {
-    baseBlack: "#505050",
-    baseWhite: "white",
-    baseRed: "#D75C5C",
-    baseLightGrey: "#AAAAAA",
-    buttonGreen:"#7bdaae",
-    buttonRed: "#f27b7f",
-    postIcon: "#8fc9eb",
-    backgroundGrey: "#fbfbfb",
-}
-
-export {color};
