+++ frontend/src/components/stream/posts/Post.jsx
@@ -1,23 +1,21 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { Card, Content, Media, Heading } from "react-bulma-components";
+import { Card, Content, Heading } from "react-bulma-components";
 
 function Post(props) {
   return (
     <div className="post">
       <Card>
         <Card.Header>
-          <Card.Header.Title>{props.post.title}</Card.Header.Title>
+          <Card.Header.Title style={{ marginLeft: "1em" }}>
+            <div>
+              <p style={{ fontWeight: "lighter" }}>@{props.post.author.displayName}</p>
+              <Heading size={5}>{props.post.title}</Heading>
+            </div>
+          </Card.Header.Title>
         </Card.Header>
         <Card.Content>
-          <Media>
-            <Media.Item>
-              <Heading size={6} style={{ marginBottom: "-0.5em" }}>
-                @{props.post.author.displayName}
-              </Heading>
-            </Media.Item>
-          </Media>
           <Content>{props.post.description}</Content>
           <p style={{ margin: "-1em 0", fontSize: "0.8em" }}>{props.post.published}</p>
         </Card.Content>
